+++ src/models/exported_qwen3_asr/tensors/decode_lm_head.py
@@ -58,7 +58,7 @@ def create_decode_lm_head(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key=None,
-                spec=TensorSpec(dtype='float32', shape=(1, 1, 1024)),
+                spec=TensorSpec(dtype='float16', shape=(1, 1, 1024)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.INPUT,
                 memory=MemoryClass.HOST_INPUT,
@@ -71,7 +71,7 @@ def create_decode_lm_head(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='linear',
-                spec=TensorSpec(dtype='float32', shape=(1, 1, 151936)),
+                spec=TensorSpec(dtype='float16', shape=(1, 1, 151936)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
